version 0.2.0 (#117)
* Increment version number to 0.1.0.9000

* add error message to use gt_update_backingfile before imputing a subs… (#91)

* add error message to use gt_update_backingfile before imputing a subset of a gt

* update CI to use dev

---------


* King threshold (#92)

* add "first" and "second" as possible arguments to kings_threshold

* additional tests

---------


* Add benchmarking (#94)

* add benchmark on hgdp
---------


* A few edits

* fix col width in article

[skip ci]

* Add reference copy of fst functions in R

* Fst fast (#96)

* faster cpp versions of loci_alt_freq, indiv_het_obs, indiv_missingness, and loci_hwe

---------


* update kinship documentation

* update documentation kings_threhsold qc_report_indiv

* Benchmark article plots (#97)

* add fst heatmap, indiv report plot, and change indiv filtering to < 1% missingness

* Add DAPC plot by region

* Fix ram reporting

---------


* create cbind method for gen_tbl

* update pkgdown

* fix pkgdown

* implement gt_add_sf

* fix lints

* better advice on cbind

[skip ci]

* Add laptop timings

* fix lints

* Simplify vignette (#98)

* simplifying names

something is wrong

* Update the metadata in a03_anole_adm.rds object loaded by vignette

* remove commented code

* reindent text

* spelling and grammar

* Add maps to a03 vignette and edit autoplot.q_matrix to handle sf geometry

* linting

---------


* Gt add sf tests (#103)

* additional gt_add_sf tests

* fix order of class returned

* update documentation

* update gt_dapc and add test

* prioritise gen_tbl class in select_loci functions

* Add rbind sf test

* add cbind test

* lint

* Group loci (#99)

* add tidy, list, matrix options to .grouped_df method for loci_alt_freq

* add tidy, list, matrix options to .grouped_df method for loci_maf

* add tidy, list, matrix options to .grouped_df method for loci_missingness

* add test

* Remove .grouped_df methods: loci_transitions, loci_transversions, ld_clump, and hwe

* Add example_gen_tibble() function and @examples to documentation

* Add description of correct use of reframe and grouped functions

* Increment version number to 0.1.0.9007

* update yml

* change example_gen_tibble to example_gt

* change variable names

* Simplify use of reframe

* Add note on pairwise verbs in vignette

* extra test

* check grouping

* add .cols tidyselect argument to loci_maf

* add .cols tidyselect argument to loci_missingness, fix message in loci_maf.grouped_df

* add .col tidyselect argument to loci_chromosomes and loci_hwe, fix loci_missingness man page

* add .col tidyselect argument to transitions, transversions, names, and add test of warning for more than two grouping variables in grouped_df methods

* change messages for polyploid alternatives

* lint

* tidy documentation

---------


* simplify a couple of thingss (#100)

* simplify a couple of thingss

* further simplifications to vignette, and additional tidyverse methods for gt objects that have sf columns

* remove group_map TODO's from loci_ functions

* update example_gt() and examples to resolve coderabbit nitpick

* make arrange.gen_tbl consistent, coderabbit nitpick

* style and lint

* update .yml

* simplify example_gt() coderabbit nitpicks and document

* quick test of example_gt

* fix comma placement styler

* lint

---------


* Projection article (#102)

* Add PCA projection article

* edit headings and start adding comments

* add a precompile

* change download to eval = FALSE

* platform independent data download

* split .figures into subfolders

* Update articles benchmark

* use cpp by default

* add laptop benchmark

---------


* Input (#93)

* draft gt_impute_xgboost

* error when imputing on partial file

* Incomplete tests for gt_impute_xgboost - error in line 78

* fix gt_impute_xgboost

* gt_impute tests, update error messages

* Add TODO

* small fixes and test for imputation

* add gt_impute_xgboost to index in .yml

---------


* adding example to test autoplot beahviour with NAs (in MAF)

* fix autoplot NA issue

* Num dem (#106)

* num and dem for Fst

* document internal function

* initial window implementation

* switch to runner to cater for bp and snp sizes

* tests for window_stats_generic

* window_pairwise_pop_fst

* initial commit for loci_pi

needs testing!

* styling and linting

* tests for loci_pi

* add num dem to test

* test pi against hierfstat

* Add pop_tajimas_d with tests

* doc fixes

* initial window_pop_tajimas_d implementation

not working, to debug

* fix windowed Tajima's D and add tests

* naming for roh function

* fix readme for dev

* Pbs (#101)

* first pbs commit

we now need to create a windowed version

* fix compile warning

* add pbs for windows with tests

* Added tests for manually checking Fst window calculations.

* clean up of windows functions

* Fix warnings

* implement codeRabbit suggestions

* More tests

* fix error spotted by codeRabbit

[skip ci]

* fix test

[skip ci]

* fix spotted by code rabbit

[skip ci]

* Minor doc fixes

[skip ci]

* spell fixes

* fix lint

* Add 3 parameters to pairwise_pop_fst: by_locus, type, and by_locus_type. Add example to windows_pairwise_fst. Fix gt_admixture spelling

* style and lint

* cpp of grouped hwe

* Add colnames and rownames to matrix option for loci functions with type = ‘matrix’

* loci_hwe grouped method

* exports cpp

* fix docs

* add 'type' to loci_pi, loci_hwe, and pairwise_pbs

* update pairwise_pop_pbs documentation

* Pairwise_pop_fst: add tidy tibble or list to by_locus_type

* Add qc_report_loci HWE group message

* Change qc_report_loci to use grouped loci_hwe method

* change warning to message and fix test

* Begin adding 'type' options to windows_ functions

* Add additional 'type' methods to pairwise_pop_pbs, add check to pairwise_pop_fst

* Document returns, change "tibble" type to "matrix" for windows_

* update pairwise_pop_pbs documentation and window_indiv_roh test

---------


* Add nwise to verbs

* fix docs

* Fst clean (#112)

* cpp for Nei and WC methods in fst
* use of nwise

* Fs stats (#113)

make pop fis and fst more efficient

* Packedancestry (#109)

* new rcpp reader for packedancestry

---------


* Tidypopgen vignette (#114)

* Add quick start vignette, fix bug in augment.gt_pca
* fix spelling
---------


* faster cpp vcf parser (#111)

* faster cpp parser for VCF

* better error message

* use additional repositories for admixtools

* try again

* readd remotes

* Benchmark heatmap (#115)

* add type = "matrix" to fix pairwise_fst bug

* Update hgdp benchmark

* Add laptop benchmark

---------


* Add pack (#116)

* test additional packages

* additional repository

* fix lint

* fix coverage

* empty lines

* bump version number

[skip ci]

* formatting

---------

+++ src/gt_grouped_summaries.cpp
@@ -30,7 +30,7 @@ ListOf<NumericMatrix> gt_grouped_summaries(Environment BM,
   for (size_t j = 0; j < m; j++) {
     for (size_t i = 0; i < n; i++) {
       double x = macc(i, j);
-      if (x>-1){
+      if (x > -1){
         freq(j, groupIds[i]) += x;
         valid_alleles(j, groupIds[i]) +=2;
         if (x==1){
@@ -39,7 +39,7 @@ ListOf<NumericMatrix> gt_grouped_summaries(Environment BM,
       }
     }
     // now for each group, divide freq by valid_alleles
-    for (int group_i = 0; group_i < ngroups; group_i++) {
+    for (size_t group_i = 0; group_i < ngroups; group_i++) {
       freq(j, group_i) = freq(j, group_i) / valid_alleles(j, group_i);
       ref_freq(j, group_i) = 1-freq(j, group_i);
       heterozygotes(j, group_i) = heterozygotes(j, group_i) / valid_alleles(j, group_i);
